Upgrade core
- Refactoring the source code.
- Fix bugs.
- Add more commands.

+++ src/Support/Installer.php
@@ -1,21 +1,12 @@
 <?php
 
-define('DS', DIRECTORY_SEPARATOR);
-
-require_once __DIR__.'/Application.php';
+namespace VhostsManager\Support;
 
 class Installer extends Application
 {
     public function __construct()
     {
-        parent::__construct(false);
-
-        if (is_file($this->paths['appDir'] . '\settings.ini') && is_file($this->paths['caCertDir'] . '\cacert.crt')) {
-            Console::breakline();
-            Console::line('Xampp vHosts Manager is already integrated into Xampp.');
-            Console::line('No need to do it again.');
-            exit;
-        }
+        parent::__construct(false, false);
     }
 
     public function install()
@@ -25,16 +16,18 @@ class Installer extends Application
         Console::breakline();
 
         Console::line('1. Register one Trusted Certificate Authority to authenticate your virtual host CSR later.');
-        Console::line('2. Grant the necessary permissions to Windows hosts file so it can be edited from this utility.');
-        Console::line('3. Improve Apache\'s "httpd-vhosts.conf" file to include vhost config files later.');
-        Console::line('4. Improve Apache\'s "httpd-ssl.conf" file to include vhost SSL config files later.');
+        Console::line('2. Improve the Apache "httpd-vhosts.conf" file to include vhost config files later.');
+        Console::line('3. Improve the Apache "httpd-ssl.conf" file to include vhost SSL config files later.');
+        Console::line('4. Grant the necessary permissions to Windows hosts file so it can be edited from this app.');
         Console::line('5. Register path into Windows Path Environment Variable, so you can call it anywhere later.');
         Console::breakline();
 
         $continue = Console::confirm('Do you agree to continue?');
 
         if (! $continue) {
-            Console::terminate(null, 1);
+            Console::breakline();
+            Console::line('The installation was canceled by user.');
+            Console::terminate(null, 2, true);
         }
 
         Console::breakline();
@@ -46,9 +39,9 @@ class Installer extends Application
         Console::breakline();
 
         $this->registerCA();
-        $this->grantPermsToWinHostsFile();
         $this->improveHttpdVhostsConfFile();
         $this->improveHttpdSslConfFile();
+        $this->grantPermsToWinHostsFile();
         $this->registerPath();
 
         Console::breakline();
@@ -60,34 +53,40 @@ class Installer extends Application
         Console::breakline();
         Console::hrline();
         Console::line('XAMPP VHOSTS MANAGER WAS INSTALLED SUCCESSFULLY.');
-        Console::line('TO START USING IT, PLEASE EXIT YOUR TERMINAL TO');
-        Console::line('DELETE TEMPORARY PROCESS ENVIRONMENT VARIABLES.');
+        Console::line('TO START USING IT, PLEASE EXIT YOUR TERMINAL TO DELETE TEMPORARY PROCESS ENVIRONMENT VARIABLES.');
     }
 
     private function askInstallConfig()
     {
-        $phpDir = dirname(PHP_BINARY);
+        $phpDir            = dirname(PHP_BINARY);
+        $maybeXamppDir     = osstyle_path(dirname($phpDir));
+        $xamppDirConfirmed = false;
 
         Console::line('First, provide the path to your Xampp directory for Xampp vHosts Manager.');
 
-        if (is_file($phpDir . '\..\xampp-control.exe')) {
-            $detectedXamppDir = realpath($phpDir . '\..\\');
-            Console::line('Xampp vHosts Manager has detected that directory "' . strtoupper($detectedXamppDir) . '" could be your Xampp directory.');
+        if (maybe_xamppdir($maybeXamppDir)) {
+            Console::line('Xampp vHosts Manager has detected that directory "' . $maybeXamppDir . '" could be your Xampp directory.');
             Console::breakline();
-            $confirmXamppDir = Console::confirm('Is that the actual path to your Xampp directory?');
+
+            $xamppDirConfirmed = Console::confirm('Is that the actual path to your Xampp directory?');
+
             Console::breakline();
         }
 
-        $xamppDir = (isset($detectedXamppDir) && $confirmXamppDir) ? $detectedXamppDir : $this->tryGetXamppDir();
+        $xamppDir = $xamppDirConfirmed ? $maybeXamppDir : $this->tryGetXamppDir();
+
         $this->setting->set('DirectoryPaths', 'Xampp', $xamppDir);
+
         $this->paths['xamppDir'] = $xamppDir;
 
-        if (is_file($xamppDir . '\apache\bin\httpd.exe')) {
+        if (maybe_apachedir($xamppDir . '\apache')) {
             $this->paths['apacheDir'] = $xamppDir . '\apache';
         } else {
             Console::line('Next, because Xampp vHosts Manager does not detect the path to the Apache directory, you need to provide it manually.');
             Console::breakline();
+
             $apacheDir = $this->tryGetApacheDir();
+
             $this->setting->set('DirectoryPaths', 'Apache', $apacheDir);
             $this->paths['apacheDir'] = $apacheDir;
         }
@@ -98,81 +97,67 @@ class Installer extends Application
             Console::terminate('Cancel the installation.', 1);
         }
 
-        $this->setting->reloadSettings();
+        $this->setting->reload();
         $this->loadAdditionalPaths();
     }
 
     private function registerCA()
     {
         $message = 'Registering Trusted CA with name "' . getenv('XVHM_OPENSSL_SUBJECT_CN') . '"...';
+
         Console::line($message, false);
 
-        $this->powerExec('"' . $this->paths['caCertGenScript'] . '"', '-w -i -n', $arrOutput, $exitCode);
+        $registered = false;
+        $generated  = $this->createdCaCert();
 
-        if ($exitCode == 0 && is_file($this->paths['caCertDir'] . '\cacert.crt')) {
+        if ($generated && $this->addedCaCertToStore()) {
+            $registered = true;
+        } elseif (!$generated) {
+            $this->powerExec('"' . $this->paths['caCertGenScript'] . '"', '-w -i -n', $arrOutput, $exitCode);
             $this->powerExec('CERTUTIL -addstore -enterprise -f -v Root "' . $this->paths['caCertDir'] . '\cacert.crt"', '-w -e -i -n', $arrOutput, $exitCode);
 
-            if ($exitCode == 0) {
-                Console::line('Successful', true, max(73 - strlen($message), 1));
-                return true;
-            }
+            $registered = true;
+        } else {
+            $this->powerExec('CERTUTIL -addstore -enterprise -f -v Root "' . $this->paths['caCertDir'] . '\cacert.crt"', '-w -e -i -n', $arrOutput, $exitCode);
+
+            $registered = true;
         }
 
-        // If generating failed
-        Console::line('Failed', true, max(77 - strlen($message), 1));
-
-        @unlink($this->paths['appDir'] . '\settings.ini');
-        $cacertGenFiles = glob($this->paths['caCertDir'] .DS. '*'); // get all file names
-        foreach ($cacertGenFiles as $file) {
-            if (is_file($file)) {
-                @unlink($file);
-            }
-        }
-
-        Console::terminate('Cancel the installation.', 1);
-    }
-
-    private function grantPermsToWinHostsFile()
-    {
-        $message = 'Granting necessary permissions to Windows host file...';
-        Console::line($message, false);
-
-        $this->powerExec('icacls "' . $this->paths['winHostsFile'] . '" /grant Users:MRXRW', '-w -i -e -n', $arrOutput, $exitCode);
-
-        if ($exitCode == 0) {
+        if ($registered) {
             Console::line('Successful', true, max(73 - strlen($message), 1));
+
             return true;
         }
 
         Console::line('Failed', true, max(77 - strlen($message), 1));
-        Console::breakline();
-        Console::line('You need set the Modify and Write permissions for group Users...');
-        Console::line('to Windows hosts file manually after installation.');
-        return false;
+        undir($this->paths['caCertDir']);
+        Console::terminate('Cancel the installation.', 1);
     }
 
     private function improveHttpdVhostsConfFile()
     {
-        $message = 'Backing up and improve Apache\'s "httpd-vhosts.conf" file...';
+        $message = 'Backing up and improving the Apache "httpd-vhosts.conf" file...';
+
         Console::line($message, false);
 
-        $backupDir = $this->paths['apacheDir'] . '\conf\extra\backup';
-        if (! is_dir($backupDir)) {
-            mkdir($backupDir, 0755, true);
-        }
-
-        // Backup httpd-vhosts.conf
+        // Vars
         $httpd_vhosts_conf = $this->paths['apacheDir'] . '\conf\extra\httpd-vhosts.conf';
-        copy($httpd_vhosts_conf, $backupDir . '\httpd-vhosts.conf');
+        $configLine        = 'IncludeOptional "' . relative_path($this->paths['apacheDir'], $this->paths['vhostConfigDir'], '/') . '/*.conf"';
+
+        // Backup
+        $this->backupFile($httpd_vhosts_conf, $this->paths['apacheDir'] . '\conf\extra\backup');
 
         // Improve httpd-vhosts.conf
-        $configLine           = 'IncludeOptional "' . $this->reduceApachePath($this->paths['vhostConfigDir'], '/') . '/*.conf"';
-        $httpd_vhosts_append  = PHP_EOL . '# Include all virtual host config files';
-        $httpd_vhosts_append .= PHP_EOL . $configLine . PHP_EOL;
-        $httpd_vhosts_updated = file_put_contents($httpd_vhosts_conf, $httpd_vhosts_append, FILE_APPEND);
+        if (line_exists($configLine, $httpd_vhosts_conf)) {
+            $httpd_vhosts_updated = true;
+        } else {
+            $httpd_vhosts_append  = PHP_EOL . '# Include all virtual host config files' . PHP_EOL . $configLine . PHP_EOL;
+            $httpd_vhosts_updated = file_put_contents($httpd_vhosts_conf, $httpd_vhosts_append, FILE_APPEND);
+        }
 
         if ($httpd_vhosts_updated) {
             Console::line('Successful', true, max(73 - strlen($message), 1));
+
             return true;
         }
 
@@ -181,31 +166,34 @@ class Installer extends Application
         Console::line('You need to add the following lines to the "' .$httpd_vhosts_conf. '" file manually after installation:');
         Console::line($configLine);
         Console::breakline();
+
         return false;
     }
 
     private function improveHttpdSslConfFile()
     {
-        $message = 'Backing up and improve Apache\'s "httpd-ssl.conf" file...';
+        $message = 'Backing up and improving the Apache "httpd-ssl.conf" file...';
+
         Console::line($message, false);
 
-        $backupDir = $this->paths['apacheDir'] . '\conf\extra\backup';
-        if (! is_dir($backupDir)) {
-            mkdir($backupDir, 0755, true);
-        }
-
-        // Backup httpd-ssl.conf
+        // Vars
         $httpd_ssl_conf = $this->paths['apacheDir'] . '\conf\extra\httpd-ssl.conf';
-        copy($httpd_ssl_conf, $backupDir . '\httpd-ssl.conf');
+        $configLine     = 'IncludeOptional "' . relative_path($this->paths['apacheDir'], $this->paths['vhostSSLConfigDir'], '/') . '/*.conf"';
+
+        // Backup
+        $this->backupFile($httpd_ssl_conf, $this->paths['apacheDir'] . '\conf\extra\backup');
 
         // Improve httpd-ssl.conf
-        $configLine        = 'IncludeOptional "' . $this->reduceApachePath($this->paths['vhostSSLConfigDir'], '/') . '/*.conf"';
-        $httpd_ssl_append  = PHP_EOL . '# Include all virtual host ssl config files';
-        $httpd_ssl_append .= PHP_EOL . $configLine . PHP_EOL;
-        $httpd_ssl_updated = file_put_contents($httpd_ssl_conf, $httpd_ssl_append, FILE_APPEND);
+        if (line_exists($configLine, $httpd_ssl_conf)) {
+            $httpd_ssl_updated = true;
+        } else {
+            $httpd_ssl_append  = PHP_EOL . '# Include all virtual host ssl config files' . PHP_EOL . $configLine . PHP_EOL;
+            $httpd_ssl_updated = file_put_contents($httpd_ssl_conf, $httpd_ssl_append, FILE_APPEND);
+        }
 
         if ($httpd_ssl_updated) {
             Console::line('Successful', true, max(73 - strlen($message), 1));
+
             return true;
         }
 
@@ -214,18 +202,43 @@ class Installer extends Application
         Console::line('You need to add the following lines to the "' .$httpd_ssl_conf. '" file manually after installation:');
         Console::line($configLine);
         Console::breakline();
+
         return false;
     }
 
-    protected function registerPath($askConfirm = true, $question = null)
+    private function grantPermsToWinHostsFile()
     {
-        $result = parent::registerPath(false);
+        $result = $this->grantPermsWinHosts(false);
+
+        if (! $result) {
+            Console::breakline();
+            Console::line('You need set the Modify and Write permissions for group Users to the Windows hosts file manually after installation.');
+        }
+    }
+
+    private function registerPath()
+    {
+        $result = $this->registerAppPath(false);
 
         if (! $result) {
             Console::breakline();
             Console::line('Don\'t worry. This does not affect the installation process.');
-            Console::line('You can register the path manually...');
-            Console::line('or use the "xvhost register_path" command after installation.');
+            Console::line('You can register the path manually or use the "xvhost register_path" command after installation.');
+        }
+    }
+
+    private function backupFile($sourceFile, $targetDir, $forceOverwrite = false)
+    {
+        if (!is_dir($targetDir)) {
+            mkdir($targetDir, 0755, true);
+        }
+
+        if (is_file($sourceFile)) {
+            $targetFile = $targetDir . '\\' . basename($sourceFile);
+
+            if ($forceOverwrite || !is_file($targetFile)) {
+                copy($sourceFile, $targetFile);
+            }
         }
     }
 
@@ -233,10 +246,10 @@ class Installer extends Application
     {
         Console::breakline();
         Console::line('At this point, the installation process is complete.');
-        Console::line('However, for Xampp vHosts Manager to work more perfectly');
-        Console::line('you should configure some more settings. You can skip this');
-        Console::line('step and configure the following via the "settings.ini" file.');
+        Console::line('However, for Xampp vHosts Manager to work more perfectly you should configure some more settings.');
+        Console::line('You can skip this step and configure the following via the "settings.ini" file.');
         Console::breakline();
+
         $configNow = Console::confirm('Would you like to do that now?');
 
         if ($configNow) {
@@ -246,15 +259,19 @@ class Installer extends Application
             Console::line('Note*: You can use the string {{host_name}} as the virtual host name placeholder.');
             Console::breakline();
 
-            $docRootSuggestion = Console::ask('Enter document root path suggestion', $this->paths['xamppDir'] . '\htdocs\{{host_name}}');
-            $this->setting->set('Suggestions', 'DocumentRoot', str_replace('/', DS, $docRootSuggestion));
+            $docRootSuggestion = $this->setting->get('Suggestions', 'DocumentRoot', $this->paths['xamppDir'] . '\htdocs\{{host_name}}');
+            $docRootSuggestion = Console::ask('Enter document root path suggestion', $docRootSuggestion);
+
+            $this->setting->set('Suggestions', 'DocumentRoot', osstyle_path($docRootSuggestion));
 
             Console::breakline();
             Console::line('[+] The second setting ---');
             Console::line('Provide the email used to propose as Admin Email each vhost creation process.');
             Console::breakline();
 
-            $adminEmailSuggestion = Console::ask('Enter admin email suggestion', 'webmaster@example.com');
+            $adminEmailSuggestion = $this->setting->get('Suggestions', 'AdminEmail', 'webmaster@example.com');
+            $adminEmailSuggestion = Console::ask('Enter admin email suggestion', $adminEmailSuggestion);
+
             $this->setting->set('Suggestions', 'AdminEmail', $adminEmailSuggestion);
 
             Console::breakline();
@@ -262,21 +279,27 @@ class Installer extends Application
             Console::line('Provide the number of records per page when listing the existing virtual hosts.');
             Console::breakline();
 
-            $recordPerPage = Console::ask('Enter the virtual hosts record per page', 2);
-            $this->setting->set('ListViewMode', 'RecordPerPage', $recordPerPage);
+            $recordPerPage = $this->setting->get('ListViewMode', 'RecordPerPage', 3);
+            $recordPerPage = Console::ask('Enter the virtual hosts record per page', $recordPerPage);
+
+            $this->setting->set('ListViewMode', 'RecordPerPage', intval($recordPerPage));
 
             Console::breakline();
+
             $message = 'Saving settings to the "settings.ini" file...';
+
             Console::line($message, false);
 
             if ($this->setting->save()) {
                 Console::line('Successful', true, max(73 - strlen($message), 1));
+
                 return true;
             }
 
             Console::line('Failed', true, max(77 - strlen($message), 1));
             Console::breakline();
             Console::line('You have to set settings through the "settings.ini" file manually.');
+
             return false;
         }
     }
@@ -284,9 +307,9 @@ class Installer extends Application
     private function tryGetXamppDir()
     {
         $xamppDir = '';
-
         $repeat = 0;
-        while (! is_file(rtrim($xamppDir, '\\/') . '\xampp-control.exe')) {
+
+        while (! maybe_xamppdir($xamppDir)) {
             if ($repeat == 4) {
                 Console::line('You have not provided correct information multiple times.');
                 Console::terminate('Cancel the installation.', 1);
@@ -296,11 +319,14 @@ class Installer extends Application
                 $xamppDir = Console::ask('Enter the path to your Xampp directory');
             } else {
                 Console::line('The path provided is not the path to the actual Xampp directory.');
+
                 $xamppDir = Console::ask('Please provide it again');
             }
 
             Console::breakline();
-            $xamppDir = str_replace('/', DS, $xamppDir);
+
+            $xamppDir = osstyle_path(rtrim($xamppDir, '/\\'));
+
             $repeat++;
         }
 
@@ -310,9 +336,9 @@ class Installer extends Application
     private function tryGetApacheDir()
     {
         $apacheDir = '';
-
         $repeat = 0;
-        while (! is_file(rtrim($apacheDir, '\\') . '\bin\httpd.exe')) {
+
+        while (! maybe_apachedir($apacheDir)) {
             if ($repeat == 4) {
                 Console::line('You have not provided correct information multiple times.');
                 Console::terminate('Cancel the installation.', 1);
@@ -322,11 +348,14 @@ class Installer extends Application
                 $apacheDir = Console::ask('Enter the path to your Apache directory');
             } else {
                 Console::line('The path provided is not the path to the actual Apache directory.');
+
                 $apacheDir = Console::ask('Please provide it again');
             }
 
             Console::breakline();
-            $apacheDir = str_replace('/', DS, $apacheDir);
+
+            $apacheDir = osstyle_path(rtrim($apacheDir, '/\\'));
+
             $repeat++;
         }
